--- a/Cargo.toml
+++ b/Cargo.toml
@@ -25,6 +25,10 @@ incremental = false
 codegen-units = 16
 rpath = false
 
+[profile.flamegraph]
+inherits = "release"
+debug = true
+
 [dev-dependencies]
 criterion = "0.5.1"