--- a/Cargo.toml
+++ b/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "huff-tree-tap"
-version = "0.1.0"
+version = "0.2.0"
 description = "Huffman Encoder and Decoder Library"
 authors = ["Alexis Lowe <alexis.lowe@chimbosonic.com>"]
 edition = "2021"