--- a/Makefile
+++ b/Makefile
@@ -1,2 +1,8 @@
 watch:
 	cargo watch -x fmt -x check -x clippy -x test
+
+flamegraph:
+	cargo flamegraph --profile flamegraph --test tests
+
+bench:
+	cargo bench