--- a/benches/huffman.rs
+++ b/benches/huffman.rs
@@ -28,7 +28,6 @@ fn criterion_benchmark(c: &mut Criterion) {
     group.finish();
 }
 
-// Encode: 650micros Decode: 1.14ms
 criterion_group!(
     name = benches;
     config = Criterion::default().warm_up_time(std::time::Duration::from_secs(10));